+++ src/ainspector-content-script/panelConstants.js
@@ -1,9 +1,0 @@
-/* panel-constants.js*/
-
-// viewId is a copy of viewId constant in panel.js
-
-export const viewId = {
-  allRules : 'all-rules',
-  ruleGroup: 'rule-group',
-  ruleResult: 'rule-result'
-};