+++ src/locale/en/common.js
@@ -49,6 +49,10 @@ export const common = {
   rulesetLevelAA:  'Levels A and AA',
   rulesetLevelAAA: 'Levels A, AA and enhanced CCR',
 
+  ruleScopeAll: "All rules",
+  ruleScopePage: "Page only rules",
+  ruleScopeWebsite: "Website only rules",
+
   rulesetFirstStep: 'First Step Rules',
   rulesetWCAG22:    'WCAG 2.2, ',
   rulesetWCAG21:    'WCAG 2.1, ',